+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Next
 * Add ability to create loadouts by selecting sets of perks
+* [#823](https://github.com/DestinyItemManager/DIM/issues/823) Added 'current' property to stores.
 
 # 3.9.2